cart_service/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy[asyncio]
asyncpg
alembic
pyjwt==2.10.1
python-dotenv
